Function truelayer_coding_challenge::server::advanced_handler

pub async fn advanced_handler(
    pokemon: PokemonResponse, 
    translation_client: impl TranslationClient, 
    cache: impl CacheWrapper<(String, TranslationType), PokemonResponse>
) -> Result<PokemonResponse, Rejection>

Filter for “advanced”, translation API requests

First determines what type of translation should be performed based on the given pokemon - as this filter always follows the basic filter, it receives a PokemonResponse and so simply queries this object. Checks cache for an existing translated response. If none, sends a request to the funtranslations API for a translation of the given Pokemon’s description. If a successful response is received, the given reponse has it’s description replaced with the translation, is cached, then returned.