Function truelayer_coding_challenge::server::router

pub fn router(
    poke_client: impl PokeClient, 
    translation_client: impl TranslationClient, 
    cache: impl CacheWrapper<(String, TranslationType), PokemonResponse>
) -> impl Filter<Extract = impl Reply, Error = Infallible> + Clone

Full router of available public API endpoints

For each path, the necessary client dependencies are injected, followed by injection of a cache reference (Moka caches are wrapped in an atomic reference count).

The “pokemon” route is simple - the PokeClient and cache are injected then the handler is invoked. The “pokemon/translated” handler effectively is an extension of the “pokemon” route - the same injection and handling as the “pokemon” route is performed, then the response handed off to be injected into the advanced handler along with a TranslationClient implementor and the cache again. In this way, the advanced handler does not need to duplicate the code to contact Pokeapi itself, and effectively reuses the basic handler to do so.